requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
djangorestframework==3.15.2
pytest==8.3.5
websockets==15.0.1
